requirements.txt
Flask
Flask-Talisman
whitenoise
python-dotenv
Flask-SQLAlchemy
Flask-Caching
gunicorn
psycopg2-binary
reportlab
requests
beautifulsoup4
